// src/api/typesGenerated.ts
export type UsersJoined = {
  user_id: string,
  is_public: boolean,
  access_code: string,
  project_id: string,
  code_path: string,
  users: User_2,
}

export type User_2 = {
  readonly id: string,
  readonly email: string,
  readonly username: string,
  readonly hashed_password: any,
  readonly created_at: string,
  readonly updated_at: string,
  readonly status: string,
  readonly rbac_roles: string[],
  readonly login_type: string,
  readonly avatar_url: string,
  readonly deleted: boolean,
  readonly last_seen_at: string,
  readonly fullname: string,
  readonly is_teacher: boolean,  
  readonly joins: ProjectsJoined[],
  readonly is_public: boolean,  
  readonly due_date: string,  
  readonly email_verified?: boolean 
}

export type ProjectsJoined = {
  users: any
  user_id: string
  is_public:boolean
  access_code: string
  project_id: string
  code_path: string
  projects: Project
}